# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: BANCO Asset Marketplace API — Cars, Real Estate, Industrial
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [INVALID_DATA, NOT_FOUND, UNAUTHORIZED, INTERNAL_ERROR, FORBIDDEN, RATE_LIMITED]
        message:
          type: string
    Meta:
      type: object
      properties:
        cursor:
          type: string
        has_next:
          type: boolean
        total:
          type: integer
...
    FeedItem:
      type: object
      description: >
        Immutable v1 contract — exactly 11 fields.
        No renames, no removals allowed in v1. Breaking changes go to /v2/*.
      required:
        - id
        - media_preview
        - price_display
        - title
        - location
        - trust_signal
        - has_video
        - is_sponsored
      properties:
        id:
          type: string
          description: Listing identifier (use for detail navigation).
        media_preview:
          type: string
        price_display:
          type: string
        installment_badge:
          type: string
          nullable: true
        title:
          type: string
        location:
          type: string
        urgency_signal:
          type: string
          nullable: true
        trust_signal:
          type: string
        smart_badge:
          type: string
          nullable: true
        has_video:
          type: boolean
        is_sponsored:
          type: boolean
        coordinates:
          $ref: "#/components/schemas/Coordinates"
          nullable: true
        best_offer_badge:
          type: string
          nullable: true
        industrial_type:
          type: string
          nullable: true
          description: Industrial sub-type (factory, warehouse, land, production_line, raw_material, machine) used for client-side category grouping. Null for non-industrial listings.
        is_active:
          type: boolean
          nullable: true
          description: True when the listing status is "active" (publicly visible and promotable). Lets owner-facing surfaces built on FeedItem (e.g. the profile grid) gate the Promote control without fetching ListingDetail. Null when status is unknown.
        whatsapp_enabled:
          type: boolean
          nullable: true
          description: True when the seller opted this listing in to WhatsApp contact. Opt-in only (default false). Lets surfaces decide whether to surface a WhatsApp action. Null when unknown.
        created_at:
          type: string
          nullable: true
          description: Listing creation timestamp (ISO 8601). Owner-facing grids (profile, "my listings") render a "Listed <date>" caption. Null only when the source row lacks it; public feeds always carry it.
        is_request:
          type: boolean
          nullable: true
          description: True when this is a buyer "request/wanted" post (looking to buy) rather than a sale listing. Surfaces render a "طلب / Wanted" badge and the requests filter keys off it. Null when unknown.
    Coordinates:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
        lng:
          type: number
...
    FinancingRequest:
      type: object
      properties:
        lead_id:
          type: string
        status:
          type: string
          enum: [new, forwarded, contacted, closed, rejected]
        listing_id:
          type: string
        listing_title:
          type: string
        category:
          type: string
          enum: [car, real_estate, industrial]
        buyer_name:
          type: string
          nullable: true
        buyer_phone:
          type: string
          nullable: true
        asset_price:
          type: string
          nullable: true
        down_payment:
          type: string
          nullable: true
        monthly_payment:
          type: string
          nullable: true
        duration_months:
          type: integer
          nullable: true
        provider_name:
          type: string
          nullable: true
        intermediary_id:
          type: string
          nullable: true
        intermediary_name:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        assigned_at:
          type: string
          nullable: true
        created_at:
          type: string
          nullable: true
        updated_at:
          type: string
          nullable: true
...
    InvoiceLineItem:
      type: object
      required: [label, amount]
      properties:
        label:
          type: string
        amount:
          type: string
    Invoice:
      type: object
      required:
        [
          id,
          invoice_number,
          amount,
          status,
          transaction_id,
          transaction_type,
          description,
          line_items,
          issued_at,
          created_at,
        ]
      properties:
        id:
          type: string
        invoice_number:
          type: string
        amount:
          type: string
        status:
          type: string
          enum: [paid, void]
        transaction_id:
          type: string
        transaction_type:
          type: string
          nullable: true
          enum:
            [
              wallet_topup,
              boost_charge,
              subscription_charge,
              lead_charge,
              refund,
              adjustment,
            ]
        description:
          type: string
          nullable: true
        line_items:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
        issued_at:
          type: string
          nullable: true
        created_at:
          type: string
          nullable: true
    BillingReportLine:
      type: object
      required: [type, total, count]
      properties:
        type:
          type: string
          enum:
            [
              wallet_topup,
              boost_charge,
              subscription_charge,
              lead_charge,
              refund,
              adjustment,
            ]
        total:
          type: string
        count:
          type: integer
    BillingReport:
      type: object
      required:
        [
          month,
          currency,
          total_charged,
          total_topped_up,
          transaction_count,
          by_type,
        ]
      properties:
        month:
          type: string
        currency:
          type: string
          enum: [EGP]
        total_charged:
          type: string
        total_topped_up:
          type: string
        transaction_count:
          type: integer
        by_type:
          type: array
          items:
            $ref: "#/components/schemas/BillingReportLine"
...
paths:
...
  /v1/feed:
    get:
      operationId: getFeed
      summary: Get adaptive scored feed
      tags: [feed]
      parameters:
        - in: query
          name: cursor
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: category
          schema:
            type: string
            enum: [car, real_estate, industrial]
        - in: query
          name: is_request
          description: >-
            Buyer request/wanted filter. Omit to include both sale listings and
            requests. true = only requests; false = only sale listings.
          schema:
            type: boolean
        - in: query
          name: industrial_type
          description: >-
            Industrial subtype filter (only meaningful with category=industrial).
            A single subtype or a comma-separated list (a whole group), e.g.
            "factory,warehouse,land". Valid subtypes: factory, warehouse, land,
            production_line, raw_material, machine.
          schema:
            type: string
        - in: query
          name: condition
          description: Filter by condition (cars). Real data via listing_attributes.condition or specs.
          schema:
            type: string
            enum: [new, used]
        - in: query
          name: payment_plan
          description: >-
            Payment engine filter. installment = any non-cash plan; bank =
            bank_finance; direct = seller_installment; islamic = sharia-compliant
            non-cash plan.
          schema:
            type: string
            enum: [installment, bank, direct, islamic]
        - in: query
          name: property_type
          description: Filter real-estate by property type (villa, apartment, penthouse, duplex, ...).
          schema:
            type: string
        - in: query
          name: finishing_type
          description: Filter real-estate by finishing level.
          schema:
            type: string
        - in: query
          name: compound
          description: Real-estate located inside a compound (specs.compound).
          schema:
            type: boolean
        - in: query
          name: furnished
          description: Furnished real-estate (specs.furnished).
          schema:
            type: boolean
        - in: query
          name: fuel_type
          description: Filter cars by fuel type (listing_attributes.fuel_type or specs).
          schema:
            type: string
            enum: [petrol, diesel, hybrid, electric, natural_gas]
        - in: query
          name: transmission
          description: Filter cars by transmission (listing_attributes.transmission or specs).
          schema:
            type: string
            enum: [manual, automatic, cvt]
        - in: query
          name: brand
          description: >-
            Filter cars by brand, matched against the English listing title
            (titles are canonical "<Brand> <Model> <Year>").
          schema:
            type: string
        - in: query
          name: model
          description: Filter cars by model, matched against the English listing title.
          schema:
            type: string
        - in: query
          name: min_year
          description: Minimum car model year (numeric specs.year).
          schema:
            type: integer
        - in: query
          name: max_year
          description: Maximum car model year (numeric specs.year).
          schema:
            type: integer
        - in: query
          name: industry
          description: Filter industrial listings by industry (listing_attributes.industry).
          schema:
            type: string
            enum: [food, beverage, plastic, textile, pharmaceutical, chemical, engineering, other]
        - in: query
          name: origin_type
          description: Filter industrial listings by origin (listing_attributes.origin_type).
          schema:
            type: string
            enum: [local, imported]
        - in: query
          name: session_id
          schema:
            type: string
      responses:
        "200":
          description: Feed items with cursor pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/FeedItem"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
...
  /v1/billing/invoices:
    get:
      operationId: listInvoices
      summary: List the caller's invoices (newest first)
      tags: [billing]
      security:
        - ClerkAuth: []
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        "200":
          description: Invoice page
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Invoice"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
  /v1/billing/invoices/{id}:
    get:
      operationId: getInvoice
      summary: Get a single invoice the caller owns
      tags: [billing]
      security:
        - ClerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Invoice"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
  /v1/billing/report:
    get:
      operationId: getBillingReport
      summary: Monthly billing summary computed from the ledger
      tags: [billing]
      security:
        - ClerkAuth: []
      parameters:
        - name: month
          in: query
          required: false
          description: Report month as YYYY-MM (UTC). Defaults to the current month.
          schema:
            type: string
      responses:
        "200":
          description: Billing report
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/BillingReport"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
...
  /v1/admin/financing/requests/{leadId}:
    patch:
      operationId: updateFinancingRequest
      summary: Bank-financing CRM — update status / intermediary assignment
      tags: [admin]
      security:
        - ClerkAuth: []
      parameters:
        - in: path
          name: leadId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [new, forwarded, contacted, closed, rejected]
                intermediary_id:
                  type: string
                  nullable: true
                notes:
                  type: string
                  nullable: true
      responses:
        "200":
          description: Updated financing request
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/FinancingRequest"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
...